Remove most TypeScript types from the core-data module in favor a simpler approach
