+++ packages/core-data/src/selectors.js
@@ -1,0 +1,985 @@
+/**
+ * External dependencies
+ */
+import createSelector from 'rememo';
+import { set, map, find, get, filter, compact } from 'lodash';
+
+/**
+ * WordPress dependencies
+ */
+import { createRegistrySelector } from '@wordpress/data';
+import { addQueryArgs } from '@wordpress/url';
+import deprecated from '@wordpress/deprecated';
+
+/**
+ * Internal dependencies
+ */
+import { STORE_NAME } from './name';
+import { getQueriedItems } from './queried-data';
+import { DEFAULT_ENTITY_KEY } from './entities';
+import { getNormalizedCommaSeparable, isRawAttribute } from './utils';
+
+/**
+ * Shared reference to an empty object for cases where it is important to avoid
+ * returning a new object reference on every invocation, as in a connected or
+ * other pure component which performs `shouldComponentUpdate` check on props.
+ * This should be used as a last resort, since the normalized data should be
+ * maintained by the reducer result in state.
+ */
+const EMPTY_OBJECT = {};
+
+/**
+ * Returns true if a request is in progress for embed preview data, or false
+ * otherwise.
+ *
+ * @param {any}    state Data state.
+ * @param {string} url   URL the preview would be for.
+ *
+ * @return {boolean} Whether a request is in progress for an embed preview.
+ */
+export const isRequestingEmbedPreview = createRegistrySelector(
+	( select ) => ( state, url ) => {
+		return select( STORE_NAME ).isResolving( 'getEmbedPreview', [ url ] );
+	}
+);
+
+/**
+ * Returns all available authors.
+ *
+ * @deprecated since 11.3. Callers should use `select( 'core' ).getUsers({ who: 'authors' })` instead.
+ *
+ * @param {any}    state Data state.
+ * @param {Object} query Optional object of query parameters to
+ *                       include with request.
+ * @return {Object[]} Authors list.
+ */
+export function getAuthors( state, query ) {
+	deprecated( "select( 'core' ).getAuthors()", {
+		since: '5.9',
+		alternative: "select( 'core' ).getUsers({ who: 'authors' })",
+	} );
+
+	const path = addQueryArgs(
+		'/wp/v2/users/?who=authors&per_page=100',
+		query
+	);
+	return getUserQueryResults( state, path );
+}
+
+/**
+ * Returns the current user.
+ *
+ * @param {any} state Data state.
+ *
+ * @return {Object|undefined} Current user object.
+ */
+export function getCurrentUser( state ) {
+	return state.currentUser;
+}
+
+/**
+ * Returns all the users returned by a query ID.
+ *
+ * @param {any} state  Data state.
+ * @param       string queryID Query ID.
+ *
+ * @return {Object[]} Users list.
+ */
+export const getUserQueryResults = createSelector(
+	( state, queryID ) => {
+		const queryResults = state.users.queries[ queryID ];
+
+		return map( queryResults, ( id ) => state.users.byId[ id ] );
+	},
+	( state, queryID ) => [ state.users.queries[ queryID ], state.users.byId ]
+);
+
+/**
+ * Returns the loaded entities for the given kind.
+ *
+ * @deprecated since WordPress 6.0. Use getEntitiesConfig instead
+ * @param {any}    state Data state.
+ * @param {string} kind  Entity kind.
+ *
+ * @return {Object} Array of entities with config matching kind.
+ */
+export function getEntitiesByKind( state, kind ) {
+	deprecated( "wp.data.select( 'core' ).getEntitiesByKind()", {
+		since: '6.0',
+		alternative: "wp.data.select( 'core' ).getEntitiesConfig()",
+	} );
+	return getEntitiesConfig( state, kind );
+}
+
+/**
+ * Returns the loaded entities for the given kind.
+ *
+ * @param {any}    state Data state.
+ * @param {string} kind  Entity kind.
+ *
+ * @return {Object[]} Array of entities with config matching kind.
+ */
+export function getEntitiesConfig( state, kind ) {
+	return filter( state.entities.config, { kind } );
+}
+
+/**
+ * Returns the entity config given its kind and name.
+ *
+ * @deprecated since WordPress 6.0. Use getEntityConfig instead
+ * @param {any}    state Data state.
+ * @param {string} kind  Entity kind.
+ * @param {string} name  Entity name.
+ *
+ * @return {Object} Entity config
+ */
+export function getEntity( state, kind, name ) {
+	deprecated( "wp.data.select( 'core' ).getEntity()", {
+		since: '6.0',
+		alternative: "wp.data.select( 'core' ).getEntityConfig()",
+	} );
+	return getEntityConfig( state, kind, name );
+}
+
+/**
+ * Returns the entity config given its kind and name.
+ *
+ * @param {any}    state Data state.
+ * @param {string} kind  Entity kind.
+ * @param {string} name  Entity name.
+ *
+ * @return {Object} Entity config
+ */
+export function getEntityConfig( state, kind, name ) {
+	return find( state.entities.config, { kind, name } );
+}
+
+/**
+ * Returns the Entity's record object by key. Returns `null` if the value is not
+ * yet received, undefined if the value entity is known to not exist, or the
+ * entity object if it exists and is received.
+ *
+ * @param {any}           state State tree
+ * @param {string}        kind  Entity kind.
+ * @param {string}        name  Entity name.
+ * @param {string|number} key   Record's key
+ * @param {Object?}       query Optional query.
+ *
+ * @return {Object|null|undefined} Record.
+ */
+export const getEntityRecord = createSelector(
+	( state, kind, name, key, query ) => {
+		const queriedState = get( state.entities.records, [
+			kind,
+			name,
+			'queriedData',
+		] );
+		if ( ! queriedState ) {
+			return undefined;
+		}
+		const context = query?.context ?? 'default';
+
+		if ( query === undefined ) {
+			// If expecting a complete item, validate that completeness.
+			if ( ! queriedState.itemIsComplete[ context ]?.[ key ] ) {
+				return undefined;
+			}
+
+			return queriedState.items[ context ][ key ];
+		}
+
+		const item = queriedState.items[ context ]?.[ key ];
+		if ( item && query._fields ) {
+			const filteredItem = {};
+			const fields = getNormalizedCommaSeparable( query._fields ) ?? [];
+			for ( let f = 0; f < fields.length; f++ ) {
+				const field = fields[ f ].split( '.' );
+				const value = get( item, field );
+				set( filteredItem, field, value );
+			}
+			return filteredItem;
+		}
+
+		return item;
+	},
+	( state, kind, name, recordId, query ) => {
+		const context = query?.context ?? 'default';
+		return [
+			get( state.entities.records, [
+				kind,
+				name,
+				'queriedData',
+				'items',
+				context,
+				recordId,
+			] ),
+			get( state.entities.records, [
+				kind,
+				name,
+				'queriedData',
+				'itemIsComplete',
+				context,
+				recordId,
+			] ),
+		];
+	}
+);
+
+/**
+ * Returns the Entity's record object by key. Doesn't trigger a resolver nor requests the entity records from the API if the entity record isn't available in the local state.
+ *
+ * @param {any}           state State tree
+ * @param {string}        kind  Entity kind.
+ * @param {string}        name  Entity name.
+ * @param {string|number} key   Record's key
+ *
+ * @return {Object|null|undefined} Record.
+ */
+export function __experimentalGetEntityRecordNoResolver(
+	state,
+	kind,
+	name,
+	key
+) {
+	return getEntityRecord( state, kind, name, key );
+}
+
+/**
+ * Returns the entity's record object by key,
+ * with its attributes mapped to their raw values.
+ *
+ * @param {any}           state State tree.
+ * @param {string}        kind  Entity kind.
+ * @param {string}        name  Entity name.
+ * @param {string|number} key   Record's key.
+ *
+ * @return {Object|null|undefined} Object with the entity's raw attributes.
+ */
+export const getRawEntityRecord = createSelector(
+	( state, kind, name, key ) => {
+		const record = getEntityRecord( state, kind, name, key );
+		return (
+			record &&
+			Object.keys( record ).reduce( ( accumulator, _key ) => {
+				if (
+					isRawAttribute( getEntityConfig( state, kind, name ), _key )
+				) {
+					// Because edits are the "raw" attribute values,
+					// we return those from record selectors to make rendering,
+					// comparisons, and joins with edits easier.
+					accumulator[ _key ] = get(
+						record[ _key ],
+						'raw',
+						record[ _key ]
+					);
+				} else {
+					accumulator[ _key ] = record[ _key ];
+				}
+				return accumulator;
+			}, {} )
+		);
+	},
+	( state, kind, name, recordId, query ) => {
+		const context = query?.context ?? 'default';
+		return [
+			state.entities.config,
+			get( state.entities.records, [
+				kind,
+				name,
+				'queriedData',
+				'items',
+				context,
+				recordId,
+			] ),
+			get( state.entities.records, [
+				kind,
+				name,
+				'queriedData',
+				'itemIsComplete',
+				context,
+				recordId,
+			] ),
+		];
+	}
+);
+
+/**
+ * Returns true if records have been received for the given set of parameters,
+ * or false otherwise.
+ *
+ * @param {any}    state State tree
+ * @param {string} kind  Entity kind.
+ * @param {string} name  Entity name.
+ * @param {Object} query Optional terms query.
+ *
+ * @return {boolean} Whether entity records have been received.
+ */
+export function hasEntityRecords( state, kind, name, query ) {
+	return Array.isArray( getEntityRecords( state, kind, name, query ) );
+}
+
+/**
+ * Returns the Entity's records.
+ *
+ * @param {any}    state State tree
+ * @param {string} kind  Entity kind.
+ * @param {string} name  Entity name.
+ * @param {Object} query Optional terms query.
+ *
+ * @return {Object[]|null} Records.
+ */
+export const getEntityRecords = ( state, kind, name, query ) => {
+	// Queried data state is prepopulated for all known entities. If this is not
+	// assigned for the given parameters, then it is known to not exist.
+	const queriedState = get( state.entities.records, [
+		kind,
+		name,
+		'queriedData',
+	] );
+	if ( ! queriedState ) {
+		return null;
+	}
+	return getQueriedItems( queriedState, query );
+};
+
+/**
+ * Returns the list of dirty entity records.
+ *
+ * @param {any} state State tree.
+ *
+ * @return {Object[]} The list of updated records
+ */
+export const __experimentalGetDirtyEntityRecords = createSelector(
+	( state ) => {
+		const {
+			entities: { records },
+		} = state;
+		const dirtyRecords = [];
+		Object.keys( records ).forEach( ( kind ) => {
+			Object.keys( records[ kind ] ).forEach( ( name ) => {
+				const primaryKeys = Object.keys(
+					records[ kind ][ name ].edits
+				).filter(
+					( primaryKey ) =>
+						// The entity record must exist (not be deleted),
+						// and it must have edits.
+						getEntityRecord( state, kind, name, primaryKey ) &&
+						hasEditsForEntityRecord( state, kind, name, primaryKey )
+				);
+
+				if ( primaryKeys.length ) {
+					const entityConfig = getEntityConfig( state, kind, name );
+					primaryKeys.forEach( ( primaryKey ) => {
+						const entityRecord = getEditedEntityRecord(
+							state,
+							kind,
+							name,
+							primaryKey
+						);
+						dirtyRecords.push( {
+							// We avoid using primaryKey because it's transformed into a string
+							// when it's used as an object key.
+							key: entityRecord[
+								entityConfig.key || DEFAULT_ENTITY_KEY
+							],
+							title:
+								entityConfig?.getTitle?.( entityRecord ) || '',
+							name,
+							kind,
+						} );
+					} );
+				}
+			} );
+		} );
+
+		return dirtyRecords;
+	},
+	( state ) => [ state.entities.records ]
+);
+
+/**
+ * Returns the list of entities currently being saved.
+ *
+ * @param {any} state State tree.
+ *
+ * @return {Object[]} The list of records being saved.
+ */
+export const __experimentalGetEntitiesBeingSaved = createSelector(
+	( state ) => {
+		const {
+			entities: { records },
+		} = state;
+		const recordsBeingSaved = [];
+		Object.keys( records ).forEach( ( kind ) => {
+			Object.keys( records[ kind ] ).forEach( ( name ) => {
+				const primaryKeys = Object.keys(
+					records[ kind ][ name ].saving
+				).filter( ( primaryKey ) =>
+					isSavingEntityRecord( state, kind, name, primaryKey )
+				);
+
+				if ( primaryKeys.length ) {
+					const entityConfig = getEntityConfig( state, kind, name );
+					primaryKeys.forEach( ( primaryKey ) => {
+						const entityRecord = getEditedEntityRecord(
+							state,
+							kind,
+							name,
+							primaryKey
+						);
+						recordsBeingSaved.push( {
+							// We avoid using primaryKey because it's transformed into a string
+							// when it's used as an object key.
+							key: entityRecord[
+								entityConfig.key || DEFAULT_ENTITY_KEY
+							],
+							title:
+								entityConfig?.getTitle?.( entityRecord ) || '',
+							name,
+							kind,
+						} );
+					} );
+				}
+			} );
+		} );
+		return recordsBeingSaved;
+	},
+	( state ) => [ state.entities.records ]
+);
+
+/**
+ * Returns the specified entity record's edits.
+ *
+ * @param {any}           state    State tree.
+ * @param {string}        kind     Entity kind.
+ * @param {string}        name     Entity name.
+ * @param {string|number} recordId Record ID.
+ *
+ * @return {Object[]|undefined} The entity record's edits.
+ */
+export function getEntityRecordEdits( state, kind, name, recordId ) {
+	return get( state.entities.records, [ kind, name, 'edits', recordId ] );
+}
+
+/**
+ * Returns the specified entity record's non transient edits.
+ *
+ * Transient edits don't create an undo level, and
+ * are not considered for change detection.
+ * They are defined in the entity's config.
+ *
+ * @param {any}           state    State tree.
+ * @param {string}        kind     Entity kind.
+ * @param {string}        name     Entity name.
+ * @param {string|number} recordId Record ID.
+ *
+ * @return {Object|undefined} The entity record's non transient edits.
+ */
+export const getEntityRecordNonTransientEdits = createSelector(
+	( state, kind, name, recordId ) => {
+		const { transientEdits } = getEntityConfig( state, kind, name ) || {};
+		const edits = getEntityRecordEdits( state, kind, name, recordId ) || {};
+		if ( ! transientEdits ) {
+			return edits;
+		}
+		return Object.keys( edits ).reduce( ( acc, key ) => {
+			if ( ! transientEdits[ key ] ) {
+				acc[ key ] = edits[ key ];
+			}
+			return acc;
+		}, {} );
+	},
+	( state, kind, name, recordId ) => [
+		state.entities.config,
+		get( state.entities.records, [ kind, name, 'edits', recordId ] ),
+	]
+);
+
+/**
+ * Returns true if the specified entity record has edits,
+ * and false otherwise.
+ *
+ * @param {any}           state    State tree.
+ * @param {string}        kind     Entity kind.
+ * @param {string}        name     Entity name.
+ * @param {string|number} recordId Record ID.
+ *
+ * @return {boolean} Whether the entity record has edits or not.
+ */
+export function hasEditsForEntityRecord( state, kind, name, recordId ) {
+	return (
+		isSavingEntityRecord( state, kind, name, recordId ) ||
+		Object.keys(
+			getEntityRecordNonTransientEdits( state, kind, name, recordId )
+		).length > 0
+	);
+}
+
+/**
+ * Returns the specified entity record, merged with its edits.
+ *
+ * @param {any}           state    State tree.
+ * @param {string}        kind     Entity kind.
+ * @param {string}        name     Entity name.
+ * @param {string|number} recordId Record ID.
+ *
+ * @return {Object|null|undefined} The entity record, merged with its edits.
+ */
+export const getEditedEntityRecord = createSelector(
+	( state, kind, name, recordId ) => ( {
+		...getRawEntityRecord( state, kind, name, recordId ),
+		...getEntityRecordEdits( state, kind, name, recordId ),
+	} ),
+	( state, kind, name, recordId, query ) => {
+		const context = query?.context ?? 'default';
+		return [
+			state.entities.config,
+			get( state.entities.records, [
+				kind,
+				name,
+				'queriedData',
+				'items',
+				context,
+				recordId,
+			] ),
+			get( state.entities.records, [
+				kind,
+				name,
+				'queriedData',
+				'itemIsComplete',
+				context,
+				recordId,
+			] ),
+			get( state.entities.records, [ kind, name, 'edits', recordId ] ),
+		];
+	}
+);
+
+/**
+ * Returns true if the specified entity record is autosaving, and false otherwise.
+ *
+ * @param {any}           state    State tree.
+ * @param {string}        kind     Entity kind.
+ * @param {string}        name     Entity name.
+ * @param {string|number} recordId Record ID.
+ *
+ * @return {boolean} Whether the entity record is autosaving or not.
+ */
+export function isAutosavingEntityRecord( state, kind, name, recordId ) {
+	const { pending, isAutosave } = get(
+		state.entities.records,
+		[ kind, name, 'saving', recordId ],
+		{ pending: false, isAutosave: false }
+	);
+	return Boolean( pending && isAutosave );
+}
+
+/**
+ * Returns true if the specified entity record is saving, and false otherwise.
+ *
+ * @param {any}           state    State tree.
+ * @param {string}        kind     Entity kind.
+ * @param {string}        name     Entity name.
+ * @param {string|number} recordId Record ID.
+ *
+ * @return {boolean} Whether the entity record is saving or not.
+ */
+export function isSavingEntityRecord( state, kind, name, recordId ) {
+	return get(
+		state.entities.records,
+		[ kind, name, 'saving', recordId, 'pending' ],
+		false
+	);
+}
+
+/**
+ * Returns true if the specified entity record is deleting, and false otherwise.
+ *
+ * @param {any}           state    State tree.
+ * @param {string}        kind     Entity kind.
+ * @param {string}        name     Entity name.
+ * @param {string|number} recordId Record ID.
+ *
+ * @return {boolean} Whether the entity record is deleting or not.
+ */
+export function isDeletingEntityRecord( state, kind, name, recordId ) {
+	return get(
+		state.entities.records,
+		[ kind, name, 'deleting', recordId, 'pending' ],
+		false
+	);
+}
+
+/**
+ * Returns the specified entity record's last save error.
+ *
+ * @param {any}           state    State tree.
+ * @param {string}        kind     Entity kind.
+ * @param {string}        name     Entity name.
+ * @param {string|number} recordId Record ID.
+ *
+ * @return {Object|undefined} The entity record's save error.
+ */
+export function getLastEntitySaveError( state, kind, name, recordId ) {
+	return get( state.entities.records, [
+		kind,
+		name,
+		'saving',
+		recordId,
+		'error',
+	] );
+}
+
+/**
+ * Returns the specified entity record's last delete error.
+ *
+ * @param {any}           state    State tree.
+ * @param {string}        kind     Entity kind.
+ * @param {string}        name     Entity name.
+ * @param {string|number} recordId Record ID.
+ *
+ * @return {Object|undefined} The entity record's save error.
+ */
+export function getLastEntityDeleteError( state, kind, name, recordId ) {
+	return get( state.entities.records, [
+		kind,
+		name,
+		'deleting',
+		recordId,
+		'error',
+	] );
+}
+
+/**
+ * Returns the current undo offset for the
+ * entity records edits history. The offset
+ * represents how many items from the end
+ * of the history stack we are at. 0 is the
+ * last edit, -1 is the second last, and so on.
+ *
+ * @param {any} state State tree.
+ *
+ * @return {number|undefined} The current undo offset.
+ */
+function getCurrentUndoOffset( state ) {
+	return state.undo.offset;
+}
+
+/**
+ * Returns the previous edit from the current undo offset
+ * for the entity records edits history, if any.
+ *
+ * @param {any} state State tree.
+ *
+ * @return {Object|null|undefined} The edit.
+ */
+export function getUndoEdit( state ) {
+	return state.undo[ state.undo.length - 2 + getCurrentUndoOffset( state ) ];
+}
+
+/**
+ * Returns the next edit from the current undo offset
+ * for the entity records edits history, if any.
+ *
+ * @param {any} state State tree.
+ *
+ * @return {Object|null|undefined} The edit.
+ */
+export function getRedoEdit( state ) {
+	return state.undo[ state.undo.length + getCurrentUndoOffset( state ) ];
+}
+
+/**
+ * Returns true if there is a previous edit from the current undo offset
+ * for the entity records edits history, and false otherwise.
+ *
+ * @param {any} state State tree.
+ *
+ * @return {boolean} Whether there is a previous edit or not.
+ */
+export function hasUndo( state ) {
+	return Boolean( getUndoEdit( state ) );
+}
+
+/**
+ * Returns true if there is a next edit from the current undo offset
+ * for the entity records edits history, and false otherwise.
+ *
+ * @param {any} state State tree.
+ *
+ * @return {boolean} Whether there is a next edit or not.
+ */
+export function hasRedo( state ) {
+	return Boolean( getRedoEdit( state ) );
+}
+
+/**
+ * Return the current theme.
+ *
+ * @param {any} state Data state.
+ *
+ * @return {string|null} The current theme.
+ */
+export function getCurrentTheme( state ) {
+	return getEntityRecord( state, 'root', 'theme', state.currentTheme );
+}
+
+/**
+ * Return the ID of the current global styles object.
+ *
+ * @param {any} state Data state.
+ *
+ * @return {string} The current global styles ID.
+ */
+export function __experimentalGetCurrentGlobalStylesId( state ) {
+	return state.currentGlobalStylesId;
+}
+
+/**
+ * Return theme supports data in the index.
+ *
+ * @param {any} state Data state.
+ *
+ * @return {any} Index data.
+ */
+export function getThemeSupports( state ) {
+	return getCurrentTheme( state )?.theme_supports ?? EMPTY_OBJECT;
+}
+
+/**
+ * Returns the embed preview for the given URL.
+ *
+ * @param {any}    state Data state.
+ * @param {string} url   Embedded URL.
+ *
+ * @return {any} Undefined if the preview has not been fetched, otherwise, the preview fetched from the embed preview API.
+ */
+export function getEmbedPreview( state, url ) {
+	return state.embedPreviews[ url ];
+}
+
+/**
+ * Determines if the returned preview is an oEmbed link fallback.
+ *
+ * WordPress can be configured to return a simple link to a URL if it is not embeddable.
+ * We need to be able to determine if a URL is embeddable or not, based on what we
+ * get back from the oEmbed preview API.
+ *
+ * @param {any}    state Data state.
+ * @param {string} url   Embedded URL.
+ *
+ * @return {boolean} Is the preview for the URL an oEmbed link fallback.
+ */
+export function isPreviewEmbedFallback( state, url ) {
+	const preview = state.embedPreviews[ url ];
+	const oEmbedLinkCheck = '<a href="' + url + '">' + url + '</a>';
+	if ( ! preview ) {
+		return false;
+	}
+	return preview.html === oEmbedLinkCheck;
+}
+
+/**
+ * Returns whether the current user can perform the given action on the given
+ * REST resource.
+ *
+ * Calling this may trigger an OPTIONS request to the REST API via the
+ * `canUser()` resolver.
+ *
+ * https://developer.wordpress.org/rest-api/reference/
+ *
+ * @param {any}           state    Data state.
+ * @param {string}        action   Action to check. One of: 'create', 'read', 'update', 'delete'.
+ * @param {string}        resource REST resource to check, e.g. 'media' or 'posts'.
+ * @param {string|number} id       Optional ID of the rest resource to check.
+ *
+ * @return {boolean| undefined} Whether or not the user can perform the action,
+ *                             or `undefined` if the OPTIONS request is still being made.
+ */
+export function canUser( state, action, resource, id ) {
+	const key = compact( [ action, resource, id ] ).join( '/' );
+	return get( state, [ 'userPermissions', key ] );
+}
+
+/**
+ * Returns whether the current user can edit the given entity.
+ *
+ * Calling this may trigger an OPTIONS request to the REST API via the
+ * `canUser()` resolver.
+ *
+ * https://developer.wordpress.org/rest-api/reference/
+ *
+ * @param {any}           state    Data state.
+ * @param {string}        kind     Entity kind.
+ * @param {string}        name     Entity name.
+ * @param {string|number} recordId Record's id.
+ * @return {boolean| undefined} Whether or not the user can edit,
+ * or `undefined` if the OPTIONS request is still being made.
+ */
+export function canUserEditEntityRecord( state, kind, name, recordId ) {
+	const entityConfig = getEntityConfig( state, kind, name );
+	if ( ! entityConfig ) {
+		return false;
+	}
+	const resource = entityConfig.__unstable_rest_base;
+
+	return canUser( state, 'update', resource, recordId );
+}
+
+/**
+ * Returns the latest autosaves for the post.
+ *
+ * May return multiple autosaves since the backend stores one autosave per
+ * author for each post.
+ *
+ * @param {any}    state    State tree.
+ * @param {string} postType The type of the parent post.
+ * @param {number} postId   The id of the parent post.
+ *
+ * @return {any[]|undefined} An array of autosaves for the post, or undefined if there is none.
+ */
+export function getAutosaves( state, postType, postId ) {
+	return state.autosaves[ postId ];
+}
+
+/**
+ * Returns the autosave for the post and author.
+ *
+ * @param {any}    state    State tree.
+ * @param {string} postType The type of the parent post.
+ * @param {number} postId   The id of the parent post.
+ * @param {number} authorId The id of the author.
+ *
+ * @return {Object|undefined} The autosave for the post and author.
+ */
+export function getAutosave( state, postType, postId, authorId ) {
+	if ( authorId === undefined ) {
+		return;
+	}
+
+	const autosaves = state.autosaves[ postId ];
+	return find( autosaves, { author: authorId } );
+}
+
+/**
+ * Returns true if the REST request for autosaves has completed.
+ *
+ * @param {any}    state    State tree.
+ * @param {string} postType The type of the parent post.
+ * @param {number} postId   The id of the parent post.
+ *
+ * @return {boolean} True if the REST request was completed. False otherwise.
+ */
+export const hasFetchedAutosaves = createRegistrySelector(
+	( select ) => ( state, postType, postId ) => {
+		return select( STORE_NAME ).hasFinishedResolution( 'getAutosaves', [
+			postType,
+			postId,
+		] );
+	}
+);
+
+/**
+ * Returns a new reference when edited values have changed. This is useful in
+ * inferring where an edit has been made between states by comparison of the
+ * return values using strict equality.
+ *
+ * @example
+ *
+ * ```
+ * const hasEditOccurred = (
+ *    getReferenceByDistinctEdits( beforeState ) !==
+ *    getReferenceByDistinctEdits( afterState )
+ * );
+ * ```
+ *
+ * @param {any} state Editor state.
+ *
+ * @return {Object} A value whose reference will change only when an edit occurs.
+ */
+export const getReferenceByDistinctEdits = createSelector(
+	// This unused state argument is listed here for the documentation generating tool (docgen).
+	( state ) => [],
+	( state ) => [
+		state.undo.length,
+		state.undo.offset,
+		state.undo.flattenedUndo,
+	]
+);
+
+/**
+ * Retrieve the frontend template used for a given link.
+ *
+ * @param {any}    state Editor state.
+ * @param {string} link  Link.
+ *
+ * @return {Object|null|undefined} The template record.
+ */
+export function __experimentalGetTemplateForLink( state, link ) {
+	const records = getEntityRecords( state, 'postType', 'wp_template', {
+		'find-template': link,
+	} );
+
+	const template = records?.length ? records[ 0 ] : null;
+	if ( template ) {
+		return getEditedEntityRecord(
+			state,
+			'postType',
+			'wp_template',
+			template.id
+		);
+	}
+	return template;
+}
+
+/**
+ * Retrieve the current theme's base global styles
+ *
+ * @param {any} state Editor state.
+ *
+ * @return {any} The Global Styles object.
+ */
+export function __experimentalGetCurrentThemeBaseGlobalStyles( state ) {
+	const currentTheme = getCurrentTheme( state );
+	if ( ! currentTheme ) {
+		return null;
+	}
+	return state.themeBaseGlobalStyles[ currentTheme.stylesheet ];
+}
+
+/**
+ * Return the ID of the current global styles object.
+ *
+ * @param {any} state Data state.
+ *
+ * @return {string|null} The current global styles ID.
+ */
+export function __experimentalGetCurrentThemeGlobalStylesVariations( state ) {
+	const currentTheme = getCurrentTheme( state );
+	if ( ! currentTheme ) {
+		return null;
+	}
+	return state.themeGlobalStyleVariations[ currentTheme.stylesheet ];
+}
+
+/**
+ * Retrieve the list of registered block patterns.
+ *
+ * @param {any} state Data state.
+ *
+ * @return {any[]} Block pattern list.
+ */
+export function getBlockPatterns( state ) {
+	return state.blockPatterns;
+}
+
+/**
+ * Retrieve the list of registered block pattern categories.
+ *
+ * @param {any} state Data state.
+ *
+ * @return {any[]} Block pattern category list.
+ */
+export function getBlockPatternCategories( state ) {
+	return state.blockPatternCategories;
+}
